moves middleware pkg one level up
